Move (wrap) layout and icon out of the UI library
First they go to the main wrap project.
In next step we will try to replace them
with ui-library components, or completely eliminate them.
