+++ ui/src/wrap/components/icons/index.ts
@@ -1,4 +1,3 @@
-export * from './WrapLogoIcon'
 export * from './LogoIconRound'
 export * from './OpenInNewIcon'
 export * from './TimesIcon'
