+++ ui/src/wrap/components/index.ts
@@ -1,6 +1,5 @@
 export * from './icons'
 export { Input as WrapInput } from './Input'
-export { Layout as WrapLayout } from './Layout'
 export { Modal as WrapModal } from './Modal'
 export * from './Spinner'
 export { ToggleButton as WrapToggleButton } from './ToggleButton'
