+++ wrap/src/App.tsx
@@ -1,9 +1,9 @@
 import { FC } from 'react'
-import { WrapLayout } from '@oasisprotocol/rose-app-ui/wrap'
+import { Layout } from './components/Layout'
 import { Web3ContextProvider } from './providers/Web3Provider'
 
 export const App: FC = () => (
   <Web3ContextProvider>
-    <WrapLayout />
+    <Layout />
   </Web3ContextProvider>
 )
