+++ wrap/src/components/Logo/WrapLogoIcon.tsx
